pyproject.toml
[tool.poetry]
name = "magic-proxy"
version = "1.0.0"
description = "Universal AI Gateway"
authors = ["Lotargo"]
license = "AGPL-3.0-only"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pyyaml = "^6.0"
aiokafka = "^0.10.0"
kafka-python = "^2.0.2"
redis = "^5.0.1"
httpx = "^0.27.0"
openai = "^1.12.0"
pydantic = "^2.5.3"
cachetools = "^5.3.2"
opentelemetry-instrumentation-fastapi = "^0.59b0"
opentelemetry-instrumentation-httpx = "^0.59b0"
opentelemetry-sdk = "^1.38.0"
python-multipart = "^0.0.9"
motor = "^3.7.1"
pytz = "^2025.1"
slowapi = "^0.1.9"
cloudinary = "^1.44.1"
python-dotenv = "^1.2.1"
beautifulsoup4 = "^4.14.3"
aiofiles = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
requests = "^2.31.0"
httpx = "^0.27.0"
locust = "^2.42.6"
py-spy = "^0.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"